# docs/api/dist/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Edge CLI",
    "version": "1.0.0",
    "description": "The `edge-cli` command line and the `edge-engine` JSON REST API, generated from the route declarations in `src/cli/engine/routes/`."
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Unix socket at ~/.edge-cli/run/<profile>/engine.sock"
    },
    {
      "url": "http://127.0.0.1:9008",
      "description": "Loopback TCP, when started with --tcp=9008. Requires the X-Edge-Token header; see securitySchemes."
    }
  ],
  "security": [
    {},
    {
      "edgeTcpToken": []
    }
  ],
...
  "paths": {
...
    "/engine/stop": {
      "post": {
        "operationId": "engineStop",
        "summary": "Stop the engine.",
        "description": "**Core call:** _none — Engine lifecycle. Internally calls `context.close()`._\n\n**Command line**\n\n```\nengine-stop\n```\n\nLogs out every session, closes the context, unlinks the socket and run-file, then exits. The engine answers before it starts tearing down, so a response is not proof the process is gone.",
        "tags": [
          "Lifecycle"
        ],
        "x-cli": {
          "command": "engine-stop",
          "flags": [],
          "extra": [],
          "custom": false,
          "preset": {}
        },
        "x-core-call": null,
        "x-core-note": "Engine lifecycle. Internally calls `context.close()`.",
        "x-source": "src/cli/engine/routes/status.ts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "Always true; a failure arrives as an error envelope."
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/engine/events": {
      "get": {
        "operationId": "engineEvents",
        "summary": "Subscribe to engine events.",
        "description": "**Core call:** _none — Engine-side fan-out; `core.log` frames carry core's onLog output._\n\n**Command line**\n\n```\nsubscribe [--session-id=<sessionId>] [--wallet-id=<walletId>] [--type=<type> …]\n```\n\nHolds a Server-Sent Events stream open until the caller disconnects or the engine closes it. Runs concurrently with one-shot calls, so a subscriber in one terminal watches what another terminal does.\n\nA live subscription holds the engine open past its idle timeout. It does not hold an account logged in: the auto-logout timer still fires, and closes any subscription scoped to that account or one of its wallets. Context-scoped subscriptions survive, because the context outlives every account.\n\nScope comes from the query string. With no `sessionId` the stream is context-scoped and nothing but the engine stopping ends it, which is what the `subscribe` command asks for. With `sessionId` it is account-scoped: a logout closes it, and the session events of *other* accounts are filtered out. `walletId` narrows it further *within* that account, for whenever a wallet-scoped event exists: no event carries a wallet scope today, so it changes nothing a caller receives, and a wallet-scoped stream gets exactly what the same stream without `walletId` would have got. It is accepted now so a caller's URL does not have to change when the first such event arrives.",
        "tags": [
          "Event stream"
        ],
        "x-cli": {
          "command": "subscribe",
          "flags": [],
          "extra": [
            {
              "name": "type",
              "kind": "repeat",
              "required": false,
              "doc": "Only these event types. Applied by the engine, so an unwanted type never crosses the socket; the client filters again for the types the engine sends regardless, like `subscription.closed`."
            }
          ],
          "custom": true,
          "preset": {},
          "notes": "Prints newline-delimited JSON and runs until interrupted. Exits 0 on SIGINT and 7 when the engine ends the stream. `subscribe` opens an unscoped stream, which only the engine stopping ends; a stream opened with `sessionId` is closed by that session logging out."
        },
        "x-core-call": null,
        "x-core-note": "Engine-side fan-out; `core.log` frames carry core's onLog output.",
        "x-source": "src/cli/engine/routes/events.ts",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": false,
            "description": "Scope the stream to one account, so a logout closes it. Omitted, the stream is context-scoped and only the engine stopping ends it.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "walletId",
            "in": "query",
            "required": false,
            "description": "With `sessionId`, narrow the stream to one wallet. No event carries a wallet scope yet, so today this changes nothing you receive — the stream still carries every account event the same `sessionId` alone would carry. Ignored on its own.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One frame per event, as `event:` then `data:` lines.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "The event name."
                    },
                    "data": {
                      "description": "Payload, shaped by the event type."
                    }
                  },
                  "required": [
                    "type",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "edgeTcpToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Edge-Token",
        "description": "The bearer token the engine mints for its loopback TCP listener, written to the `0600` run file at ~/.edge-cli/run/<profile>/engine.json as `tcpToken`. Required on the TCP transport and ignored on the unix socket. A TCP request that also carries `Origin`, or a `Host` other than the bound address, is refused with 403 whatever token it presents."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "description": "Every failure, on both transports.",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "status": {
                "type": "number"
              },
              "details": {}
            },
            "required": [
              "code",
              "message",
              "status"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}